chore: separating styles from some components
