+++ src/pages/Play/PlayPage.styles.tsx
@@ -1,0 +1,16 @@
+import styled from 'styled-components'
+
+export const Container = styled.section`
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  margin: 0 auto;
+
+  max-width: 450px;
+  vertical-align: center;
+  .bgio-client {
+    height: 100%;
+    width: 100%;
+  }
+`
